Move SCC line class to a specific file
